requirements.txt
annotated-types==0.7.0
anyio==4.6.0
babel==2.16.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.16.8
et-xmlfile==1.1.0
fastapi==0.110.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.0
idna==3.10
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.1
nodeenv==1.9.1
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
prisma==0.15.0
prisma-client==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
pywin32==307
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.37.2
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.29.0
websockets==12.0
